Convert into a proper package
